rename dictionary pkg to symbol tables pkg
